+++ simplq/src/components/Layout/Layout.module.scss
